Prevent duplicate records for an export file
Because the tax-info export method creates a record in the
`tax_info_export` table exery time it runs there would be duplicates
if it ran more than once on the same batch. This commit will prevent
that by first trying to look up the export recocord for the current
batch id.